% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice-theme.R
\name{farbe}
\alias{farbe}
\alias{likert_col}
\title{Color palettes}
\usage{
farbe(
  type = c("pirat", "ggplot", "dark", "dunkel", "pastel", "hell", "cb", "color.blinde",
    "sex", "sex.mf", "bw", "grays", "likert", "Reds", "Blues", "Greens", "Blues",
    "Greys", "Oranges", "Purples"),
  n = 5,
  ...
)

likert_col(n = 5, name = "RdBl", middle = mean(1:n), middle.color = "gray90")
}
\arguments{
\item{type}{"pirat", "ggplot",
"dark", "pastel","cb", "sex","sex.mf","brewer","bw","grays"}

\item{n}{Number of different colors in the palette, minimum 3, maximum depending on palette}

\item{...}{an Likert_col name, middle, middle.color}

\item{name}{A palette name from the lists below "RdBl"  ist   RdBl = c("Reds", "Blues")}

\item{middle, middle.color}{reference   "gray65"}
}
\value{
character-string
}
\description{
Hex-6-Code
}
\details{
#FF0000   Maximum an rot und kein greun oder blau

#00FF00  nur gruen und kein rot oder blau

#0000FF  nur blau und kein rot oder gruen


Hex-8-Code

 Leuchtkraft und Zahlenwert

Hex-Wert	 	Leuchtkraft (Prozent)
00	 	0 Prozent
40	 	25 Prozent
80	 	50 Prozent
C0	 	75 Prozent
FF	 	100 Prozent
}
\examples{

farbe()

 par(mfrow=c(2,3))
barplot(cbind(1:5, rep(3,5)), horiz = TRUE, col=farbe("likert", 5, name="RdBl"))
barplot(cbind(1:3, rep(3,3)), horiz = TRUE, col=farbe("likert", 3, name="RdBl"))
barplot(cbind(1:8, rep(3,8)), horiz = TRUE, col=farbe("likert", 8, name="RdBl"))

barplot(cbind(1:5, rep(3,5)), horiz = TRUE, col=farbe("pirat", 5))
barplot(cbind(1:3, rep(3,3)), horiz = TRUE, col=farbe("Greens", 3))
barplot(cbind(1:8, rep(3,8)), horiz = TRUE, col=farbe("color.blinde", 8))




farbe("likert")
stp25settings:::likert_col(5 )
#"#FC9272"    "#FEE0D2"     "gray90"    "#DEEBF7"    "#9ECAE1"


farbe("likert", n=4,  name="GrRd", middle=2)
stp25settings:::likert_col(4 , "GrRd", middle=2)
#"#A1D99B"     "gray90"    "#FEE0D2"    "#FC9272"

}
