% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_lang}
\alias{get_lang}
\alias{set_lang}
\alias{delet_lang}
\alias{Names2Language}
\title{Automatisches Uebersetzen}
\usage{
get_lang(x = NULL)

set_lang(...)

delet_lang(...)

Names2Language(x, replace)
}
\arguments{
\item{x}{name nin get_lang}

\item{...}{alles an set_lang}

\item{replace}{in Names2Language}
}
\description{
Names2Language() wurde von Output benutzt um die De/En Uebersaetzung durchzufuehren.
}
