% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice-theme.R
\name{set_lattice}
\alias{set_lattice}
\alias{reset_lattice}
\alias{standard_theme}
\alias{ggplot_theme}
\alias{bw_theme}
\title{Graphical Parameters for Lattice}
\usage{
set_lattice(theme = bw_theme())

reset_lattice()

standard_theme()

ggplot_theme(
  col = c("#00BA38", "#00BFC4", "#619CFF", "#F564E3", "#F8766D", "#B79F00"),
  col.bar = NULL,
  pch = 15:18,
  lty = 1:3,
  cex = 1,
  cex.main = 1.2,
  cex.axis = 0.8,
  cex.xlab = 1,
  cex.ylab = 1,
  cex.add = 0.8,
  strip.background.col = c("grey80", "grey70", "grey60"),
  ...
)

bw_theme(
  col = grDevices::grey.colors(7, start = 0.3, end = 0.9),
  col.n = 5,
  col.bar = "grey50",
  pch = 15:18,
  lty = 1:3,
  lwd = c(2, rep(1, 6)),
  cex = 1,
  cex.main = 1.2,
  cex.axis = 0.8,
  cex.xlab = 1,
  cex.ylab = 1,
  cex.add = 0.8,
  cex.symbol = 0.8,
  strip.background.col = "transparent",
  strip.border.col = "#000000",
  strip.text.col = "#000000",
  strip.text.cex = NULL,
  strip.text.lineheight = NULL,
  box = NULL,
  ...
)
}
\arguments{
\item{theme}{default= bw_theme()}

\item{col, col.bar, col.n}{Farben}

\item{pch, lty, lwd}{Symbole}

\item{cex, cex.main, cex.axis, cex.xlab, cex.ylab, cex.add, cex.symbol}{Schrift und Symbol-Groese}

\item{strip.background.col, strip.border.col, strip.text.col, strip.text.cex, strip.text.lineheight}{Strip}

\item{...}{alles was direkt an trellis.par.set uebergeben wird}

\item{box}{Box}
}
\value{
invisible liste der lattice-settings
}
\description{
ggplot lattice-Theme

Black and White lattice-Theme
}
\examples{


axis.padding   <- lattice.getOption("axis.padding")
layout.heights <- lattice.getOption("layout.heights")
layout.widths  <- lattice.getOption("layout.widths")

#' Abstand der inneren Labels in effect::plot() (ist ja ein lattice-plot)
lattice.options(axis.padding =list(numeric=0.3))

#' Abstand oben, unten, links, rechts
lattice.options(
  layout.widths =
    list(
      left.padding=list(x=.4),
      right.padding=list(x=1)
    )
  )

#' reset
lattice.options(
  layout.widths = layout.widths,
  layout.heights =layout.heights,
  axis.padding =axis.padding
  )


# require(stp25settings)
require(lattice)
#lattice::trellis.par.set(standard_theme())
lattice::trellis.par.set(bw_theme(farbe()))
barchart(yield ~ variety | site, data = barley,
         groups = year, layout = c(1,6), stack = TRUE,
         auto.key = list(space = "right"),
         ylab = "Barley Yield (bushels/acre)",
         scales = list(x = list(rot = 45)))


 #lattice::trellis.par.set(
 #  effects::effectsTheme(
 #
 #    strip.background = list(col = "transparent"),
 #    strip.shingle = list(col = "black"),
 #    clip = list(strip = "off"),
 #    superpose.line = list(lwd = c(2, rep(1, 6))),
 #    col = grey.colors(7, start = 0.3, end = 0.9)
 #    ))

}
