% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-opt.R
\name{set_opt}
\alias{set_opt}
\alias{get_opt}
\alias{default_stp25}
\title{Settings}
\usage{
set_opt(...)

get_opt(name = NULL, type = NULL)

default_stp25(type, name)
}
\arguments{
\item{name, type}{spezifikation was gesucht ist 'prozent', und 'digits'}

\item{output, fig_folder, html_folder, center}{wird über Projekt() eingestellt "html"}

\item{caption}{logical TRUE => include N in caption}

\item{sep_element, brackets}{globale symbole}

\item{table}{list(
stubhead  = "Item",
measure.name.m = "m",
measure.name.sd = "sd",
measure.name.total ="Total",
measure.name.statistics ="Statistics")}

\item{prozent}{list(
digits = 0,
lead.zero = TRUE,
style = 1,
percentage_str = "%",
null_percent_sign = ".",
include_name= "",
include_level_multi=TRUE,
exclude)}

\item{mean}{list(
digits = 2,
lead.zero = TRUE,
plusmin_str = intToUtf8(177),
style = 1,
seperator = ", ",
include_name= "(mean)" )}

\item{median}{list(
digits = 2,
lead.zero = TRUE,
seperator = ", ",
style = "IQR",
include_name="(median)" )}

\item{Fstat}{list(
digits = 2,
lead.zero = TRUE,
include.statistic=TRUE )}

\item{r}{list(
digits = 2,
lead.zero = FALSE)}

\item{r2}{list(
digits = 2,
lead.zero = FALSE)}

\item{p}{list(
digits = 3,
lead.zero = FALSE,
stars.value = c(0.001, 0.01, 0.05),
stars.symbols = c("***", "**", "*"),
with.stars = FALSE,
mark.sig = NULL )}
}
\value{
string
}
\description{
Settings
}
\examples{

set_opt(
percent = list(
  style = 2,
  null_percent_sign =  ' . ',
  digits = 0,
  percentage_str = " \%",
  include_name = ", n (\%)"
),

mean = list(style = 1,
            include_name = ", mean (sd)"),

median = list(style = 3,
              include_name = ", median (range)"),

Fstat =list(include.statistic=FALSE),

table = list(
  stubhead = "Items",
  measure.name.m = "Mittelwert/Prozent",
  measure.name.total = "Summe",
  measure.name.statistics = "Statistik"
)
)
 get_opt()
}
